+++ packages/block-editor/src/components/rich-text/utils.jsx
@@ -1,6 +1,3 @@
-import { renderToString } from '@wordpress/element';
-import { createBlock } from '@wordpress/blocks';
-
 export function addActiveFormats( value, activeFormats ) {
 	if ( activeFormats?.length ) {
 		let index = value.formats.length;
@@ -29,6 +26,10 @@ export function getMultilineTag( multiline ) {
 	return multiline === true ? 'p' : multiline;
 }
 
+export function isEmpty( value ) {
+	return ! value || value.length === 0;
+}
+
 export function getAllowedFormats( { allowedFormats, disableFormats } ) {
 	if ( disableFormats ) {
 		return getAllowedFormats.EMPTY_ARRAY;
@@ -38,17 +39,3 @@ export function getAllowedFormats( { allowedFormats, disableFormats } ) {
 }
 
 getAllowedFormats.EMPTY_ARRAY = [];
-
-/**
- * Creates a link from pasted URL.
- * Creates a paragraph block containing a link to the URL, and calls `onReplace`.
- *
- * @param {string}   url       The URL that could not be embedded.
- * @param {Function} onReplace Function to call with the created fallback block.
- */
-export function createLinkInParagraph( url, onReplace ) {
-	const link = <a href={ url }>{ url }</a>;
-	onReplace(
-		createBlock( 'core/paragraph', { content: renderToString( link ) } )
-	);
-}