+++ packages/components/src/card/card-body/hook.ts
@@ -1,8 +1,7 @@
-import { useMemo } from '@wordpress/element';
+import clsx from 'clsx';
 import type { WordPressComponentProps } from '../../context';
 import { useContextSystem } from '../../context';
-import * as styles from '../styles';
-import { useCx } from '../../utils/hooks/use-cx';
+import styles from '../style.module.scss';
 import type { BodyProps } from '../types';
 import { getPaddingBySize } from '../get-padding-by-size';
 
@@ -17,20 +16,16 @@ export function useCardBody(
 		...otherProps
 	} = useContextSystem( props, 'CardBody' );
 
-	const cx = useCx();
-
-	const classes = useMemo(
-		() =>
-			cx(
-				styles.getCardBodyStyles( { isScrollable } ),
-				styles.borderRadius,
-				getPaddingBySize( size ),
-				isShady && styles.shady,
-				// This classname is added for legacy compatibility reasons.
-				'components-card__body',
-				className
-			),
-		[ className, cx, isScrollable, isShady, size ]
+	const classes = clsx(
+		styles.body,
+		getPaddingBySize( size ),
+		{
+			[ styles[ 'is-scrollable' ] ]: isScrollable,
+			[ styles[ 'is-shady' ] ]: isShady,
+		},
+		// This classname is added for legacy compatibility reasons.
+		'components-card__body',
+		className
 	);
 
 	return {