+++ packages/widget-dashboard/CHANGELOG.md
@@ -30,7 +30,7 @@
 
 ### Bug Fixes
 
--   Widget error boundary: Announce only the error message instead of the entire notice. ([#82737](https://github.com/WordPress/gutenberg/pull/82737))
+-   Widget error boundary: Explicitly announce widget errors with polite priority. ([#82737](https://github.com/WordPress/gutenberg/pull/82737))
 
 
 ## 0.7.0 (2026-09-10)